Replace hyphen to underscore.
The hyphen in id and class replace to underscore.
In id because the hyphen is not valide in name variable in Python and as
the ids are usable as a variable it was not compatible.
And as I replacing in id I also wanted to replace in class, I would use the same logic for id and class.